Add HTML snippet for film screenings
The snippet keeps us in-line with the DRY principle but it will likely
have to change if it's going to be used for non-Agile venues. Or at
least, it needs to be renamed to indicate it's specifically for those
venues.

As the Agile venues don't use the same JSON data to link back to their
websites, the external_id attribute in the AgileEvent is replaced with
an info_link attribute that points to the individual film's page in the
Agile web page instead of the venue's actual website.