Add a module that pulls ticketing info from the MBC RSS feed
The RSS feed has some drawbacks: each showtime is listed as its own
entry so in this commit, a class is instantiated for each showtime
regardless if it's for the same film. A future commit needs to implement
a way to combine showtimes for the same film into one instance.

Due to the above issue, the film snippet cannot be used to render the
data on the webpage itself. For now, the code is added into the index
template with minor modifications in order to match the way the data
is presented in the RSS feed (as opposed to the Agile feed).