Add password resetting so new users can set their password
With this commit, user "registration" is possible in a specific way: a
SQL INSERT statement is executed on the target database where only the
email is inserted and the active field is set to True. The new user
can then go to the login page and follow the 'Forgot your password?'
procedure to set a new password.

Registration is done this way because we do not actually want anyone
registering themselves without admin approval. A future feature should
lock the (currently disabled) Flask-Security registration feature behind
the Superuser view to make future additions easier (instead of a SQL
insert).

The initial user, however, must still be created via the SQL statement
when the app is first deployed.

The roles and & users tables were pluralized they'd be in line with the
other table names. In addition, it fixes a conflict with the Postgresql
reserved user table.